finished first version

+++ FIFO_register.h
@@ -5,23 +5,23 @@
 class FifoReg{
     private:
         int length;
-        float* regist;
+        int* regist;
 
     public:
         FifoReg(int len) {
             this->length=len;
-            regist = new float[length];
+            regist = new int[length];
         }
 
-        void FifoReg_shift(float new_val) {
+        void FifoReg_shift(int new_val) {
             for (int i=length-1; i>1; i--) {
                 regist[i] = regist[i-1];
             }
              regist[0] = new_val;
         }
 
-        double FifoReg_shift_and_m_av(float new_val) {
-            volatile double average=0.000000;
+        int FifoReg_shift_and_m_av(int new_val) {
+            volatile int average=0;
             for (int i=length-1; i>=1; i--) {
                 regist[i] = regist[i-1];
                 average += regist[i];
@@ -36,7 +36,7 @@ class FifoReg{
 
         void FifoReg_print() {
             for (int i=length-1; i>=0; i--) {
-                printf("%6f ",regist[i]);
+                printf("%6d ",regist[i]);
 
             }
             printf("\n ");
